network: Move server to pkg/network/dhcp/server
