+++ pkg/network/dhcp/server/server_suite_test.go
@@ -1,4 +1,4 @@
-package dhcp
+package server
 
 import (
 	"testing"
