+++ pkg/network/driver/BUILD.bazel
@@ -10,11 +10,11 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/network/cache:go_default_library",
+        "//pkg/network/dhcp/server:go_default_library",
         "//pkg/network/link:go_default_library",
         "//pkg/util/sysctl:go_default_library",
         "//pkg/virt-handler/selinux:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter:go_default_library",
-        "//pkg/virt-launcher/virtwrap/network/dhcp:go_default_library",
         "//pkg/virt-launcher/virtwrap/network/dhcpv6:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
