--- pkg/virt-launcher/virtwrap/network/dhcp/BUILD.bazel
@@ -1,32 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["dhcp.go"],
-    importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/network/dhcp",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/network/dhcp/server:go_default_library",
-        "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
-        "//staging/src/kubevirt.io/client-go/log:go_default_library",
-        "//vendor/github.com/krolaw/dhcp4:go_default_library",
-        "//vendor/github.com/vishvananda/netlink:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = [
-        "dhcp_suite_test.go",
-        "dhcp_test.go",
-    ],
-    embed = [":go_default_library"],
-    deps = [
-        "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
-        "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
-        "//vendor/github.com/krolaw/dhcp4:go_default_library",
-        "//vendor/github.com/onsi/ginkgo:go_default_library",
-        "//vendor/github.com/onsi/gomega:go_default_library",
-        "//vendor/github.com/vishvananda/netlink:go_default_library",
-    ],
-)
